:recycle: Rename desktop main class
